"t" is float not int

+++ covid19.c
@@ -657,7 +657,7 @@ int * initialize_infections(int * initial_infections, float * tau, int * infecte
 }
 
 
-void segment_population(int* num_sus, int* num_infectious, int* num_hosp, int* num_icu, int* infected, int* infectious, int* sus_list, int* hosp_list, int * hosp_pop, int * icu_pop, int* icu_list, float* tau, int population, int t) {
+void segment_population(int* num_sus, int* num_infectious, int* num_hosp, int* num_icu, int* infected, int* infectious, int* sus_list, int* hosp_list, int * hosp_pop, int * icu_pop, int* icu_list, float* tau, int population, float t) {
 
 	*num_sus=0;
 	*num_infectious=0;
